requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
SQLAlchemy==2.0.29
pg8000==1.31.2
python-dotenv==1.0.1
pydantic==2.7.1
email-validator==2.1.1